+++ apollo-server-app/src/middlewares/AuthTokenHandlerMiddleware.js
@@ -1,8 +1,6 @@
 import jwt from 'jsonwebtoken';
 import "regenerator-runtime/runtime";
 import fetch from 'node-fetch';
-import dotenv from 'dotenv';
-dotenv.config();
 
 //Refresh Token, Post Request
 const requestRefreshToken =  async (token) => {
@@ -19,8 +17,9 @@ const requestRefreshToken =  async (token) => {
 //Set Header response helper
 export const setHeadersResponseMiddleware = (argument, response) => {
   const { context } = argument;
-  const { refresh } = context;
-  if(context.exp_status){
+  const { refresh, exp_status } = context.extra;
+  if(exp_status){
+    argument.response.http.headers.set('Access-Control-Expose-Headers', 'x-token, x-refresh-token');
     argument.response.http.headers.set('x-token', refresh.access);
     argument.response.http.headers.set('x-refresh-token', refresh.refresh);
   }
@@ -29,12 +28,11 @@ export const setHeadersResponseMiddleware = (argument, response) => {
 
 //Refresh middleware
 export const RefreshTokenMiddleware = async (request) => {
-  // Get the user token from the headers.
-  const token = request.headers['x-token'] || '';
-  const refreshToken = request.headers['x-refresh-token'] || '';
-  const currentTime = new Date().getTime() / 1000;
-
   try {
+    // Get the user token from the headers.
+    const token = request.headers['x-token'] || '';
+    const refreshToken = request.headers['x-refresh-token'] || '';
+    const currentTime = new Date().getTime() / 1000;
     //Decode token
     const tokenDecode = jwt.decode(token);
     //Chek exp